Renames the tableHeadersService utility to tableHeaders.ts and moved to table-tool module under utils
