+++ src/explore-education-statistics-frontend/src/modules/table-tool/utils/tableHeaders.ts
